add access control to books controller

+++ src/app/controllers/books.controller.js
@@ -38,30 +38,12 @@ module.exports = {
     },
 
     insert: async(req, res, next) => {
-
-        if (!req.files || Object.keys(req.files).length === 0) {
-            return res.status(400).send('No files were uploaded.');
-        }
-
-        const file = req.files.file;
-        const fileSize = file.data.length;
-        const ext = path.extname(file.name);
-        const fileName = file.md5 + ext;
-        const url = `${req.protocol}://${req.get("host")}/images/${fileName}`;
-        const allowedType = ['.png', '.jpg', '.jpeg'];
-
-        if (!allowedType.includes(ext.toLowerCase())) return res.status(422).json({ msg: "Invalid Images" });
-        if (fileSize > 5000000) return res.status(422).json({ msg: "Image must be less than 5 MB" });
-
-        file.mv(`./public/images/${fileName}`, async(err) => {
-            if (err) return res.status(500).json({ msg: err.message });
-            try {
-                await Product.create({ name: name, image: fileName, url: url });
-                res.status(201).json({ msg: "Product Created Successfuly" });
-            } catch (error) {
-                console.log(error.message);
-            }
-        })
+        if (res.locals.admin == false)
+            return res.status(403).json({
+                success: false,
+                code: 403,
+                message: "Forbidden resource"
+            })
         try {
             const response = await book.create({
                 idBuku: req.body.idBuku,
@@ -88,6 +70,12 @@ module.exports = {
     },
 
     updateId: async(req, res, next) => {
+        if (res.locals.admin == false)
+            return res.status(403).json({
+                success: false,
+                code: 403,
+                message: "Forbidden resource"
+            })
         try {
             const idBuku = req.params.id
             const response = await book.update({
@@ -129,6 +117,12 @@ module.exports = {
     },
 
     remove: async(req, res, next) => {
+        if (res.locals.admin == false)
+            return res.status(403).json({
+                success: false,
+                code: 403,
+                message: "Forbidden resource"
+            })
         try {
             await book.destroy({
                 where: {
